requirements.txt
hydra-core==1.1.0
pytorch-lightning==1.3.8
test-tube==0.7.5
timm==0.4.12
dalib==0.2